requirements.txt
langchain
flask
sentence-transformers
pypdf
python-dotenv
langchain-pinecone
langchain-openai
langchain-community
langchain-google-genai
ipykernel
psycopg2>=2.9.11
psycopg2-binary>=2.9.11
bcrypt>=5.0.0
decouple>=0.0.7
fastapi[all]>=0.128.0
jwt>=1.4.0
pyjwt[crypto]>=2.10.1
python-decouple>=3.8
sqlalchemy>=2.0.46

-e .